requirements.txt
facebook_sdk==3.1.0
Pillow==9.0.0
requests==2.26.0
schedule==1.1.0
